ml_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-auth
earthengine-api
shapely
geojson

